fine-tuning accuracy

+++ includes/image.h
@@ -5,6 +5,6 @@ typedef struct image_struct_t {
 	int height;
 	int number_of_pixels;
 	int channels_per_pixel;
-	long *channels;
+	long long *channels;
 	struct image_struct_t *next;
 } Image;
